Tradingview telebot/requirements.txt
python-telegram-bot
python-binance
python-kucoin
pandas
numpy 